requirements.txt
python-dotenv
tqdm
pandas
playwright